fix transaction issue when exception occurs
This should close ctron/package-drone#141 which also affects
the current master of Eclipse Package Drone.

When something goes whoops the StorageModelProvider does not have a
chance to clean up afterwards.

Now the model gets the chance to close itself and this is being used
to rollback any pending transactions on the BlobStore and CacheStore.