#10: P2 repository merges plugins with different versions
When the extracted feature information is transformed into the
InstallableUnit XML structure the requirements are mapped to a key.
However the key must not be unique, since it does not (as the Feature
Information) contain the version. Therefore this mapping processing
will actually merge the "bundle+version" identifier to one "bundle"
identifier.

This patch uses a simple list now to collect requirement and provides
as this reflects the actual structure of the Installation Unit.