#28: refactor the cleanup aspect to work with the new trigger system
This change will refactor common parts of the cleanup aspect to a
common bundle and will then implement both the existing cleanup aspect
and the new cleanup processor based on this common base.

The new cleanup processors works exactly the same as the cleanup aspect
but now it is possible to have multiple instances of the cleanup
functionality, in different configuration and being able to execute
those at different times.