#72: allow controlling 32/64 bit mode of RPM creation
This change allows to control whether 32 or 64bit integers will be used
for inodes and devices. 64 bit integers causes issues with other RPM
versions (4.4 and before).

It is now possible to specify whether to force 32 bit, 64 bit or let
the builder evaluate what is needed. If 64bit integers are needed but
the build is forced to 32 bit, then an exception will be thrown.